+++ test/e2e/framework/log/logger.go
@@ -14,14 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package log should be removed after switching to use core framework log.
+// Package log will be removed after switching to use core framework log.
+// Do not make further changes here!
 package log
 
 import (
 	"fmt"
-	"regexp"
-	"runtime/debug"
-	"strings"
 	"time"
 
 	"github.com/onsi/ginkgo"
@@ -51,9 +49,8 @@ func Failf(format string, args ...interface{}) {
 // (for example, for call chain f -> g -> FailfWithOffset(1, ...) error would be logged for "f").
 func FailfWithOffset(offset int, format string, args ...interface{}) {
 	msg := fmt.Sprintf(format, args...)
-	skip := offset + 1
-	log("FAIL", "%s\n\nFull Stack Trace\n%s", msg, PrunedStack(skip))
-	ginkgowrapper.Fail(nowStamp()+": "+msg, skip)
+	log("FAIL", msg)
+	ginkgowrapper.Fail(nowStamp()+": "+msg, 1+offset)
 }
 
 // Fail is a replacement for ginkgo.Fail which logs the problem as it occurs
@@ -63,42 +60,6 @@ func Fail(msg string, callerSkip ...int) {
 	if len(callerSkip) > 0 {
 		skip += callerSkip[0]
 	}
-	log("FAIL", "%s\n\nFull Stack Trace\n%s", msg, PrunedStack(skip))
+	log("FAIL", msg)
 	ginkgowrapper.Fail(nowStamp()+": "+msg, skip)
-}
-
-var codeFilterRE = regexp.MustCompile(`/github.com/onsi/ginkgo/`)
-
-// PrunedStack is a wrapper around debug.Stack() that removes information
-// about the current goroutine and optionally skips some of the initial stack entries.
-// With skip == 0, the returned stack will start with the caller of PruneStack.
-// From the remaining entries it automatically filters out useless ones like
-// entries coming from Ginkgo.
-//
-// This is a copy of https://github.com/onsi/ginkgo/blob/f90f37d87fa6b1dd9625e2b1e83c23ffae3de228/internal/codelocation/code_location.go#L25
-func PrunedStack(skip int) string {
-	fullStackTrace := string(debug.Stack())
-	stack := strings.Split(fullStackTrace, "\n")
-	// Ensure that the even entries are the method names and the
-	// the odd entries the source code information.
-	if len(stack) > 0 && strings.HasPrefix(stack[0], "goroutine ") {
-		// Ignore "goroutine 29 [running]:" line.
-		stack = stack[1:]
-	}
-	// The "+2" is for skipping over:
-	// - runtime/debug.Stack()
-	// - PrunedStack()
-	skip += 2
-	if len(stack) > 2*skip {
-		stack = stack[2*skip:]
-	}
-	prunedStack := []string{}
-	for i := 0; i < len(stack)/2; i++ {
-		// We filter out based on the source code file name.
-		if !codeFilterRE.Match([]byte(stack[i*2+1])) {
-			prunedStack = append(prunedStack, stack[i*2])
-			prunedStack = append(prunedStack, stack[i*2+1])
-		}
-	}
-	return strings.Join(prunedStack, "\n")
 }
